% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_gig.R
\name{gig}
\alias{gig}
\title{Create a function to generate generalised inverse gaussian family}
\usage{
gig(M2 = 45, M3 = 32)
}
\arguments{
\item{M2}{Numerical of length 1.}

\item{M3}{Numerical of length 1.}
}
\value{
A function to create a family object with parameters:
\itemize{
\item \code{a}: Numerical of length 1.
\item \code{b}: Numerical of length 1.
\item \code{alpha}: Numerical of length 1.
}
See \url{https://en.wikipedia.org/wiki/Generalized_inverse_Gaussian_distribution}.
\code{alpha} is \eqn{p} in the page.
}
\description{
A function to create "family" objects is returned. The "family" object is to be used in the \code{family}
argument in \code{\link{ruin_prob_ls}}. It contains all essential information to calculate the
ruin probability, e.g. calculation of \eqn{\Theta_{f,k}} for various functions \eqn{f}.
}
\details{
Appendix A3 in article "Finite-time ruin probabilities using bivariate Laguerre series" is used.
The sum to infinity terms in (A18), (A19) are truncated at \code{M2} and \code{M3} (\code{M2>M3}).

Ideally, large \code{M2} and \code{M3} should be used. However, due to limit in numerical precision in R
(long double in C, i.e. 64bit) calculation of \eqn{\Theta_{f,k}} for large \eqn{k} is unstable. As a result,
\code{M2} and \code{M3} has to be moderate.

Computer programs with high precision arithmetic built-in (e.g. Mathematica or Matlab) should be used for
reliable results.
}
\examples{

library(pruin)

a <- 5
b <- 0.02
alpha <- 2.5

family <- gig()(a=a,b=b,alpha=alpha) # a list object

family[c("name","par","mean")] # extract some information

}
