% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_LS.R
\name{ruin_prob_ls}
\alias{ruin_prob_ls}
\title{Computing finite-time ruin probability for compound poisson process with selected claims distribution}
\usage{
ruin_prob_ls(
  u,
  t,
  c,
  lambda,
  family,
  M = 20,
  psi_u = NULL,
  u_scale = 1,
  t_scale = 1,
  check = FALSE,
  include_error = FALSE
)
}
\arguments{
\item{u}{Numerical of length 1. Initial surplus.}

\item{t}{Numerical of length 1. Terminal time.}

\item{c}{Numerical of length 1. Premium rate.}

\item{lambda}{Numerical of length 1. Claim arrival rate ("lambda" parameter of the Poisson process).}

\item{family}{An object returned by the family constructors, e.g. \code{exponential()()},
\code{comb_exponential()()}, \code{gig()()}, \code{truncated_normal()()}  and \code{weibull()()}.}

\item{M}{Numerical of length 1. Default to be 20. How many Laguerre terms are used.}

\item{psi_u}{Numerical of length 1 if not \code{NULL}. Whether a user-supplied \eqn{\psi(u,t=\infty)} is used.
For some distribution (e.g. exponential) this value can be computed analytically. It generally have negligible
impact on the output.}

\item{u_scale}{Numerical of length 1. Whether a monetary scale is applied. it would have a huge impact on the ouput.
Recommend to use \code{\link{uscale_search}} to determine.}

\item{t_scale}{Numerical of length 1. Whether a time scale is applied. It generally have negligible impact on the
output.}

\item{check}{Logical. Default to be \code{FALSE}. Whether to print some sensible checks.}

\item{include_error}{Logical. Default to be \code{FALSE}. Whether to include the approximation error of
\eqn{\psi(u=0,t=\infty)}.}
}
\value{
Numerical of length 1 if \code{include_error} is \code{FALSE}. Otherwise a list of length 2, with
elements "res" and "err".
}
\description{
Computing finite-time  ruin probability with initial surplus \eqn{u} and terminal time \eqn{t}, denoted
\eqn{\psi(u,t)}.
Method described in "Finite-time ruin probabilities using bivariate Laguerre series" is used. For numerical
stability, argument \code{u_scale} has to be tuned via \code{\link{uscale_search}}. In addition, \code{u} cannot be
too large.
}
\details{
The net profit condition must be satisfied, i.e. \code{c-lambda*family$mean>0}.
}
\examples{

library(pruin)

ruin_prob_ls(u=10,t=2,c=1.1,lambda=1,family=exponential()(beta=1),u_scale = 0.6)

}
