% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_truncated_normal.R
\name{truncated_normal}
\alias{truncated_normal}
\title{Create a function to generate truncated normal family}
\usage{
truncated_normal(M2 = 45, M3 = 32)
}
\arguments{
\item{M2}{Numerical of length 1.}

\item{M3}{Numerical of length 1.}
}
\value{
A function to create a family object with parameters:
\itemize{
\item \code{mu}: Numerical of length 1.
\item \code{sigma}: Numerical of length 1.
}
}
\description{
A function to create "family" objects is returned. The "family" object is to be used in the \code{family}
argument in \code{\link{ruin_prob_ls}}. It contains all essential information to calculate the
ruin probability, e.g. calculation of \eqn{\Theta_{f,k}} for various functions \eqn{f}.
}
\details{
Appendix A3 in article "Finite-time ruin probabilities using bivariate Laguerre series" is used.
The sum to infinity terms in (A18), (A19) are truncated at \code{M2} and \code{M3} (\code{M2>M3}).

Ideally, large \code{M2} and \code{M3} should be used. However, due to limit in numerical precision in R
(long double in C, i.e. 64bit) calculation of \eqn{\Theta_{f,k}} for large \eqn{k} is unstable. As a result,
\code{M2} and \code{M3} has to be moderate.

Computer programs with high precision arithmetic built-in (e.g. Mathematica or Matlab) should be used for
reliable results.
}
\examples{

library(pruin)

mu <- 1
sigma <- 1

family <- truncated_normal()(mu=mu,sigma=sigma) # a list object

family[c("name","par","mean")] # extract some information

}
